implement side bar
