+++ nav_bar/src/Sidebar.css
@@ -1,0 +1,30 @@
+.side-nav-container{
+  position: fixed;
+  height: 100vh;
+  width: 200px;
+  top: 40px;
+  bottom: 100px;
+  box-shadow: inset -1px 0 0 #e5e5e5;
+  z-index: 600;
+  background-color: #F1F1F1; 
+}
+
+.subnav-menu-links{
+    display: flex;
+    color: #000;
+    padding: 10px 16px 10px 16px;
+    text-decoration: none;
+    flex-direction: column;    
+}
+
+.subnav-menu-links:hover{
+  background-color: #dedede; 
+}
+
+.menu-items{ 
+  list-style-type: none;
+  margin: 0;
+  padding: 0; 
+  border: 1px; 
+  overflow: auto;
+}
